Simplify DynamoDBRelation.
Here we switch from implementing `PrunedFilteredScan` to `PrunedScan` because
the filter implementation was incomplete. The filtered trait will be added
back when fully implemented.

We also switch to a somewhat convoluted process of turning items to json, then
parsing the json to get the Row's. This enables support for all data types in
the short term and performance can be improved over time.